requirements.txt
tensorflow==2.13.1
keras==2.13.1
fastapi==0.103.2
gradio==3.50.2
requests==2.31.0
zipfile36==0.1.3
scipy==1.13.0
scikit-learn==1.4.2
pillow==10.3.0